# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Test Management Platform API
  description: |
    Comprehensive API for the Test Management Platform with Observer/Orchestrator architecture.
    
    This API provides endpoints for:
    - Authentication and user management
    - Test execution orchestration
    - Git repository management
    - Webhook integrations for CI/CD systems
    
    ## Authentication
    The API uses session-based authentication. Users must login to receive a session cookie
    that will be used for subsequent authenticated requests.
    
    ## Architecture
    The platform follows an Observer/Orchestrator pattern where the API coordinates
    external CI/CD systems rather than executing tests directly.
  version: 1.0.0
  contact:
    name: Test Management Platform
    url: http://localhost:3000
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /tests/executions:
    get:
      tags:
        - Test Orchestration
      summary: List test executions
      description: Get paginated list of test executions
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
      responses:
        '200':
          description: List of test executions
          content:
            application/json:
              schema:
                type: object
                properties:
                  executions:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestExecution'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /tests/run:
    post:
      tags:
        - Test Orchestration
      summary: Orchestrate test execution
      description: Trigger test execution via external CI/CD system
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - repositoryId
                - branch
              properties:
                repositoryId:
                  type: integer
                  example: 1
                branch:
                  type: string
                  example: main
                testSuite:
                  type: string
                  example: e2e
                environment:
                  type: string
                  example: staging
                priority:
                  type: string
                  enum: [low, normal, high, urgent]
                  default: normal
      responses:
        '201':
          description: Test execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  executionId:
                    type: string
                    example: exec_123456
                  message:
                    type: string
                    example: Test execution orchestrated successfully
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using HTTP cookies
  schemas:
...
    TestExecution:
      type: object
      properties:
        id:
          type: string
          example: exec_123456
        repositoryId:
          type: integer
          example: 1
        branch:
          type: string
          example: main
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        testSuite:
          type: string
          example: e2e
        environment:
          type: string
          example: staging
        priority:
          type: string
          enum: [low, normal, high, urgent]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: number
          description: Duration in seconds
        results:
          type: object
          properties:
            total:
              type: integer
            passed:
              type: integer
            failed:
              type: integer
            skipped:
              type: integer
        artifacts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              url:
                type: string
              type:
                type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
...